backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=2.0.0
openai>=1.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
